pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "nb41"
dynamic = ["version"]
description = "NIOBIUM: Nadia's Image Occlusion Booster Is UnManned - A CLI tool for extracting text and image-occlusion-style notes from images and PDFs for Anki"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0"}
authors = [
    {name = "Agah Karakuzu", email = "agahkarakuzu@gmail.com"},
]
keywords = ["anki", "ocr", "image-occlusion", "pdf", "education", "flashcards"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "easyocr",
    "pillow",
    "numpy<2.0",
    "PyMuPDF",
    "tqdm",
    "requests",
    "rich",
    "genanki",
    "anthropic>=0.39.0",
    "pymupdf4llm",
    "pyyaml",
    "rich-argparse",
]

[project.urls]
Homepage = "https://github.com/agahkarakuzu/niobium"
Repository = "https://github.com/agahkarakuzu/niobium"
Issues = "https://github.com/agahkarakuzu/niobium/issues"

[project.scripts]
niobium = "anki_niobium.cli:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["anki_niobium"]
